backend/requirements.txt
fastapi==0.105.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
qdrant-client==1.9.0
openai==1.3.8
requests==2.31.0
pdfplumber==0.10.1
PyPDF2==3.0.1
numpy>=1.20.0
pyjwt>=2.4.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.5
cryptography>=3.4.0
google-generativeai==0.4.1